+++ noodles-bgzf/src/async/reader/builder.rs
@@ -1,3 +1,5 @@
+use std::num::NonZeroUsize;
+
 use futures::TryStreamExt;
 use tokio::io::AsyncRead;
 
@@ -7,7 +9,7 @@ use crate::Block;
 /// An async BGZF reader builder.
 pub struct Builder<R> {
     inner: R,
-    worker_count: Option<usize>,
+    worker_count: Option<NonZeroUsize>,
 }
 
 impl<R> Builder<R>
@@ -28,11 +30,16 @@ where
     /// # Examples
     ///
     /// ```
+    /// use std::num::NonZeroUsize;
+    ///
     /// use noodles_bgzf as bgzf;
+    ///
     /// let data = [];
-    /// let builder = bgzf::AsyncReader::builder(&data[..]).set_worker_count(8);
+    /// let worker_count = NonZeroUsize::try_from(1)?;
+    /// let builder = bgzf::AsyncReader::builder(&data[..]).set_worker_count(worker_count);
+    /// # Ok::<_, std::num::TryFromIntError>(())
     /// ```
-    pub fn set_worker_count(mut self, worker_count: usize) -> Self {
+    pub fn set_worker_count(mut self, worker_count: NonZeroUsize) -> Self {
         self.worker_count = Some(worker_count);
         self
     }
@@ -47,10 +54,13 @@ where
     /// let reader = bgzf::AsyncReader::builder(&data[..]).build();
     /// ```
     pub fn build(self) -> Reader<R> {
-        let worker_count = self.worker_count.unwrap_or_else(num_cpus::get);
+        let worker_count = self.worker_count.unwrap_or_else(|| {
+            // SAFETY: `num_cpus::get` is guaranteed to be non-zero.
+            NonZeroUsize::new(num_cpus::get()).unwrap()
+        });
 
         Reader {
-            stream: Some(Inflater::new(self.inner).try_buffered(worker_count)),
+            stream: Some(Inflater::new(self.inner).try_buffered(worker_count.get())),
             block: Block::default(),
             position: 0,
             worker_count,